feat: fetching orders

+++ database/migrations/2023_06_01_132938_create_menus_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('name');
             $table->text('description');
             $table->string('category');
-            $table->decimal('price', 8,2);
+            $table->float('price');
             $table->string('status');
             $table->string('image')->nullable();
             $table->timestamps();
